checks: do not use fake passthru
Exposing passhtru attributes as the derivation attributes is done by `stdenv.mkDerivation` so if we just add them to the passthru attribute of the resulting derivation attrset, they will be only available inside the passthru attribute. We can just append those attributes to the derivation attrset directly, it won’t affect the derivation either way.